refactor(parser): Move CustomExporter to exporters directory
